remove SPI and rebuild load interface
