+++ iotdb-core/node-commons/src/main/java/org/apache/iotdb/commons/service/metric/MetricService.java
@@ -26,9 +26,15 @@ import org.apache.iotdb.commons.service.JMXService;
 import org.apache.iotdb.commons.service.ServiceType;
 import org.apache.iotdb.metrics.AbstractMetricService;
 import org.apache.iotdb.metrics.config.ReloadLevel;
+import org.apache.iotdb.metrics.core.IoTDBMetricManager;
+import org.apache.iotdb.metrics.core.reporter.IoTDBJmxReporter;
 import org.apache.iotdb.metrics.metricsets.IMetricSet;
+import org.apache.iotdb.metrics.reporter.Reporter;
 import org.apache.iotdb.metrics.reporter.iotdb.IoTDBInternalMemoryReporter;
 import org.apache.iotdb.metrics.reporter.iotdb.IoTDBInternalReporter;
+import org.apache.iotdb.metrics.reporter.iotdb.IoTDBSessionReporter;
+import org.apache.iotdb.metrics.reporter.prometheus.PrometheusReporter;
+import org.apache.iotdb.metrics.utils.ReporterType;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -42,6 +48,41 @@ public class MetricService extends AbstractMetricService implements MetricServic
 
   private MetricService() {
     // empty constructor
+  }
+
+  @Override
+  protected void loadManager() {
+    metricManager = IoTDBMetricManager.getInstance();
+  }
+
+  @Override
+  protected void loadReporter() {
+    LOGGER.info("Load metric reporters, type: {}", METRIC_CONFIG.getMetricReporterList());
+    compositeReporter.clearReporter();
+    if (METRIC_CONFIG.getMetricReporterList() == null) {
+      return;
+    }
+    for (ReporterType reporterType : METRIC_CONFIG.getMetricReporterList()) {
+      Reporter reporter = null;
+      switch (reporterType) {
+        case JMX:
+          reporter = IoTDBJmxReporter.getInstance();
+          break;
+        case PROMETHEUS:
+          reporter = new PrometheusReporter(metricManager);
+          break;
+        case IOTDB:
+          reporter = new IoTDBSessionReporter(metricManager);
+          break;
+        default:
+          break;
+      }
+      if (reporter == null) {
+        LOGGER.warn("Failed to load reporter which type is {}", reporterType);
+        continue;
+      }
+      compositeReporter.addReporter(reporter);
+    }
   }
 
   @Override
